chore: separete MainNavigation, remove code duplication, put helpers in common
